requirements.txt
# Django>=2
Pillow>=6.1.0

